os.analysis: Fix default thread state when syscall are not traced
The current behavior is to assign sys_clone to a new thread where the parent's
system call is unknown. When system calls are not traced, the sys_clone exit
event does not occurs, and therefore the thread stays in syscall state, which
is incorrect. In fact, a thread should not enter the system call state when no
syscalls are traced.

For this reason, this patch does not speculate about the state of the
parent, and the child inherit whatever value is set for the parent.

Change-Id: Ic4a3ea8cab7d487a8effe3df282e134292ff8ec1
Signed-off-by: Francis Giraldeau <francis.giraldeau@gmail.com>
Reviewed-on: https://git.eclipse.org/r/68076
Reviewed-by: Hudson CI
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Tested-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>