lttng: make Critical path listen to thread selected signal instead of selection listener
This patch greatly simplifies the critical path parameter provider.
Now it will listen to a TmfThreadSelectedSignal instead of listenning to an
action on the control flow view. This will decouple the critical path from the
control flow view, instead allowing anything generating the signal the opportunity
to run a critical path analysis on a given thread and trace.

Change-Id: Ifb7f1fe2d5f81a8127ceab93b262cc9c0bbdd0e5
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/68099
Reviewed-by: Hudson CI
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Tested-by: Genevieve Bastien <gbastien+lttng@versatic.net>