tmf: add events field analysis requirement class and test
This class allows to specify required events fields for one given event
type.

This class is useful for CTF, where optional event contexts are provided
as event fields.

Moreover, this will be useful for custom txt and xml trace types, where
for each event the event type is the same but the event fields might
be different.

Change-Id: I445976ea7b93ee95be887504bc223912c2cab298
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/68233
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>