lttng: Fix intermittent MemoryUsageViewTest failure
It is possible that a thread gets added to the memory usage chart
before the process name is determined. Once the process name is
determined, a new series is added to the chart but the old one is not
deleted. Not only this fails the test because the chart contains too
many series, but it shows the wrong information to the user.

To reproduce the issue, you can add a Thread.sleep(5000) in
UstMemoryStateProvider around line 222, before the process name gets
added to the state system.

Change-Id: I95631758e0fdffbc094a3adb923b7019841a4a60
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/68312
Reviewed-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>