ss: no longer have a strings section in the HTNodes
The state systems nodes had a string section to allow for faster seeks
on the state system file if the entire interval was not loaded to memory.
This is never the case and it adds an extra integer to each non-integer state
as well as extra (but well tested) logic.

This patch makes the state sytem store each interval sequentially on a
given node.

Expect minimal performance difference, however, the size of the state system
may shrink a bit (approx 5%).

Change-Id: I01bc8594b7944fad97b6c9b715b848b6afa39913
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>
Reviewed-on: https://git.eclipse.org/r/68404
Reviewed-by: Hudson CI