control: Bug 489857: Fix live mode support for LTTng 2.4
Live streaming was added in 2.4.x. Since we can't retrieve whether
a session is live or not from a 2.4.x node, we copy the information
over from the session information at session creation time.

Note that this information gets lost when the user presses the refresh
button or restarts or Trace Compass.

For LTTng 2.5.0 and later it's possible to retrieve this information.
So we don't need copy this at creation time. This will make sure that
the session is recognized as live after a Trace Compass restart or a
refresh.

Change-Id: I4f07e0502de9c858a614946640305fbbcc8c27f6
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/68685
Reviewed-by: Hudson CI
Reviewed-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Tested-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/69000
Reviewed-on: https://git.eclipse.org/r/69074