tmf: Introduce "External Analyses" and "Reports" project elements
These will contain the upcoming LTTng-Analyses wrappers, along
with the reports they will produce.

"External Analyses" will be refered to as "on-demand analyses" in
the code, as this is what this sub-tree will eventually become.
We can use this name for now to better differentiate with the
"standard" analyses, which are still shown in the Project View.

Change-Id: I9ef5296f22a121a7b8bc85c4681a43e7a5ee5054
Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>
Reviewed-on: https://git.eclipse.org/r/70717
Reviewed-by: Hudson CI
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>