swtbot: Stabilize use of SWTBotUtils.maximizeTable()
When Ctrl+M is pressed, the resize and layout of the workbench part does
not necessarily occur immediately. For tests that depend on the part
being maximized, the method should block until the resize has occurred.

So far, this issue was only seen when SWTKeyboardStrategy is used.
SWT strategy posts events to SWT and this differs from the AWT strategy.
But the solution in this patch offers an additional safety that is
beneficial to all keyboard strategies.

Change-Id: I3afac9314a82c319064accca1fce00c1dbeb893e
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/69075
Tested-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-by: Hudson CI
Reviewed-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>