os.linux: Add per cpu thread 0 modeling
The current model handle thread 0 as a normal thread which is incorrect
since it is the idle or swapper thread and can be active concurently on
multiple CPUs.

This commit adds this concept to the state system and the control flow
view by using a special thread attribute name of "0_X" where X is the
CPU number.

Change-Id: I476927be378482751ee523fac77bbf4d260da54a
Signed-off-by: Michael Jeanson <mjeanson@efficios.com>
Reviewed-on: https://git.eclipse.org/r/69089
Reviewed-by: Hudson CI
Reviewed-by: Alexandre Montplaisir <alexmonthy@efficios.com>
Tested-by: Alexandre Montplaisir <alexmonthy@efficios.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>