lttng.ust: Implement a symbol provider for LTTng-UST traces
This new symbol provider will make use of the "Debug Info"
analysis, which means it will only be available for traces
taken with LTTng-UST >= 2.8.

It will call into the existing BinaryCallsite and SourceCallsite
aspects to get the function name of a given address.

The provider's configuration page will allow specifying the
"root directory" of a target image. This is useful in cases
where a trace is taken on a remote target, and an image of that
target is available locally.

Since paths are found directly in the trace, this means that a
trace with the proper information taken on the same system
should show the function names in the Callstack View without
the need for the user to configure anything!

Possible future improvements:
- Split the symbol provider extension point in core and ui parts
- Merge the FileOffsetMapper logic with the core parts of the
  symbol provider.
- Move the "current path prefix" configuration from the trace
  object to the core symbol provider.

References bug 484042

Change-Id: Ia346c1eab2b54a54ddf672c7ffafd0655a47ee34
Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>
Reviewed-on: https://git.eclipse.org/r/69973
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>