tmf.ui: Make TmfEventsTable#PackColumns return a value
This allows extending classes to know whether the pack actually happened
or if it was fast returned. The quick return shows that the columns have
already been packed. If the method does pack the columns, we know that it
was done the first time (on a new trace open) or after. This can allow
extending classes to have a hook to a location where the columns have
been initialized as they are not yet in the consutructor or in "init".

Change-Id: Ia4f3ab76e144d937bb5a8eb5e7597c9e99b3114f
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/69469
Reviewed-by: Hudson CI
Reviewed-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Tested-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>