linux.ui: Add Threads Presentation to View Menu of the CFV
This patch adds a threads presentation item to the view menu
(the triangular menu in the top right corner of the view) of
the control flow view, in order to select which presentation
mode is needed.

At the moment, the options offered are:
 * Hierarchical
 * Flat

Others may come later such as clamping by time...

These menus are the first step towards grouping threads by
affinity.

Change-Id: I65a71ad4a5eb6489f668f8b5806f12702f43edbc
Signed-off-by: Mahdi Zolnouri <mzolnouri@yahoo.com>
Signed-off-by: Samuel Gagnon <samuel.gagnon92@gmail.com>
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/68231
Reviewed-by: Hudson CI