tmf: Bug 490400: Leaking widgets due to incorrect cleanup in dispose()
Subclasses of Widget now perform their cleanup in a DisposeListener.

Disposing any child Control of a Composite is handled by the platform
and is therefore removed from the cleanup handling.

Subclasses of TmfComponent that are associated with a Control now
dispose their Control when TmfComponent.dispose() is called, and perform
their cleanup in a DisposeListener, including calling super.dispose().
The DisposeListener is attached to the Control.

Change-Id: Ie6bf7b18b8fad6a0fc0a683209a92fa39bbc8907
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/70057
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Hudson CI