lttng.ust: Add the build-ID to the key of cache of addr2line calls
Before calling addr2line, we should also verify that the buildId
we got from the trace matches the file on disk we are about to
look at.

This buildId is already present in the UstDebugInfoBinaryFile
objects returned by the analysis, it is just a matter of
passing it down to the FileOffsetMapper call.

This way we can use it as part of the key for the cache of
calls, so that eventually calling it on different binaries
yields separate calls.

The actual verification is not done at the moment, as this
would require calling a separate process (like "eu-readelf").
There is not much gain in doing it yet anyway, because we only
look for one possible file (the one at the expected path). If
eventually the analysis is extended to look through several
different files for a given path - using separate debug symbol
files for examples - then the surrounding code won't have to
be modified.

Change-Id: I11f42a4295c88f66ad4a685198e30181daf0a1ba
Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>
Reviewed-on: https://git.eclipse.org/r/70306
Reviewed-by: Hudson CI
Reviewed-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Tested-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>