tmf: Support for context-sensitive menu in Time Graph combos
Up to now, there were specific context-sensitive menus in Call Stack
and Control Flow view using the right mouse button on the tree.

This commit adds a generic support for context-sensitive menus to the
AbstractTimeGaph views when using Time Graph combos. It allows
extending classes to add both programmatically and declaratively
(using the plug-in.xml) menu entries to the context-sensitive menu.

The implementation of the Call Stack view and Control Flow view have
been adjusted to use this new API.

Change-Id: Ic0f0971ac32767446d0f555315f6ecbf0aeba09a
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/70136
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Hudson CI