tmf: Support for context-sensitive menu in Time Graph viewers
Up to now, there was a specific context-sensitive menu in the
Resources Flow view using the right mouse button on the namespace.

This commit adds a generic support for context-sensitive menus to the
AbstractTimeGraph views when using Time Graph viewers alone. It allows
extending classes to add both programmatically and declaratively
(using the plug-in.xml) menu entries to the context-sensitive menu.

The implementation of the Resources view have been adjusted to use this
new API

Change-Id: Ic9f3ff08fda27ffabc91d6054fe460b8fc764835
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/70140
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>