analysis: Remove zoomed event list handling from Critical Path View
The Critical Path View does not use a zoomed event list so there is no
need to fake it. Its event list is built by the HorizontalLinksVisitor,
and contains all time events regardless of the resolution.

The zoomed event list was built only once for the full entry range and
was therefore strictly equal to the full event list. It was not more
detailed (since the full event list is already at the most detailed
level) and it was not a subset of the event list for the current time
range, since the zoom thread never updated it, due to the
CriticalPathBaseEntry having no child entries, preventing any recursion
into the deeper entries.

Change-Id: I469ef9314dcbcc835a5ef03c93d9ca11c5f7eb70
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/70184
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Tested-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Reviewed-by: Francis Giraldeau <francis.giraldeau@gmail.com>