tmf: Update AbstractTimeGraphView and TimeGraphEntry API
AbstractTimeGraphView.buildEventList() is renamed to buildEntryList()
which is more representative of its implementation. The Javadoc of this
method and getEventList() is improved to better explain their use.

TimeGraphEntry methods setParent(), getParent(), getChildren() and
addChild(...) are changed to restrict the entries to TimeGraphEntry.
This avoids instanceof checks in AbstractTimeGraphView implementations,
and avoids unexpected behavior in case a child or parent entry is not of
type TimeGraphEntry.

Change-Id: Ibf1c3f3fad467ab7ff215fd2552bb3350f00ef7e
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/70185
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>