lttng: Make schema validation optional
With this change, the schema validation is disabled by default except
for tests. The validation can be enabled easily as a Tracing option
in the launch configuration. To be able to enable this by default for
tests, it is also exposed in a system variable which could also be
used on the command-line.

Change-Id: I492f2877a7e066e14176a209159da17cea83c3f1
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/70199
Reviewed-by: Hudson CI
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>