tmf: Open XML editor when clicking Edit in XML analysis manager
This change adds an Edit button the the XML analysis manager dialog
which opens a simple XML editor (from WTP). Dependencies are included
in the RCP on a per-plugin basis but work is being done to make a
feature containing the minimum set of plugins for a simple XML editor
(see bug 480141). Once this feature is available, the RCP will be
updated to use that instead, for better maintainability.

Change-Id: I6fef003210b3457815de0ce15b22e81be2eb2f9f
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/70200
Reviewed-by: Hudson CI