Add some debug output to TestRefreshTextTrace
This test has been failing intermittently but my suspicion is that
it fails when the server is under heavy load and the trace fails to
index in 20 seconds. To verify this, this change adds some output
as the wait condition is evaluated. This test didn't already log the
SWTBot output so this means that even with this change, it's not too
verbose.

Change-Id: Ic6da7b19c45faa5345b2118c902101db6e18276b
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/70329
Reviewed-by: Hudson CI
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>