ss: Add serialization logic to state values
Right now the HTInterval object takes care of state value
serialization. This is a bad separation of responsibilities,
the state value itself should be the master of its destiny.

Implement state value serialization to and from byte arrays.
In a second step, the HT can be moved to use that logic instead
of its own.

Change-Id: I1d3379e0a4d7d82816164ae214838462ab28c5fd
Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>
Reviewed-on: https://git.eclipse.org/r/70337
Reviewed-by: Hudson CI
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Tested-by: Genevieve Bastien <gbastien+lttng@versatic.net>