tmf: Add/update '-', '+' and '*' key handling in TimeGraphViewer
The handling of '-'and '+' key have been updated when the mouse is over
the namespace. Before they were used for time zooming regardless of the
mouse position. The handling of '*' key has been added when the mouse
is over the namespace.

When the mouse is over the namespace, pressing the
- key '-' will collapse recursively selected entry
- key '+' will expand selected entry
- key '*' will expand selected entry to the level with at least one
  collapsed entry

This patch is to align the keystroke handling to the TimeGraphCombo.

Change-Id: I964168bd1aa388430775055a39f625beaffb5985
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/70583
Reviewed-by: Hudson CI
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>