tmf: Increase timeout for indexing in TestRefreshTextTrace
The additional output added recently to this test confirmed that
during the test, the trace is being slowly indexed without being
stuck and ends up timing out. This change significantly increases
the timeout (20 secs to 5 mins) in order to allow more indexing
time when the machine running the test is under heavy load.

Change-Id: I8e76ced07a5a09b6bd14a7b11d12c925c1ce3df2
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/70581
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>