tmf: Bug 491714: Fix population of XmlTimeGraphView for experiments
With this patch the XmlTimeGraphView is populated with the content of
all traces part of an experiment.

When rebuilding the event list the view cache (event list, filters,
view context) is cleared. This is important because right now
there is only one instance of XmlTimeGraphView per trace even if there
are more than one XML analyses and views defined.

Change-Id: I535cace0089147bea4584c5af8b7a0c57157bd36
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/70681
Reviewed-by: Hudson CI
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>