ss: Rework the HTNode cache
Reimplement the HTNode cache as a Guava LoadingCache. This
nicely abstracts the loading logic, and will also allow
making cache accesses multi-threaded.

This also changes the log points from CacheHit/CacheMiss to
CacheLookup/CacheMiss, since the "cache hit" branch is now
entirely part of the Guava library, so we cannot instrument
it. Same information is available, it will just have to be
computed differently.

Change-Id: I267008c69f9d6f4ada0257dee45b2a79902b8c84
Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>
Reviewed-on: https://git.eclipse.org/r/70713
Reviewed-by: Hudson CI
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>