Skip irrelevant modules when using Sonarqube 5.x
sonar.skippedModules is depracated and will not work in more recent
versions of Sonarqube. We can define sonar.skip in respective modules
instead. Both properties can be defined without problem. To preserve
the pom-less build, a new profile is activated in the presence of the
sonar.skip.properties file. So each module we do not want to analyse
with maven have to contain this file.

Change-Id: I6bf8115540744e48a9431f63ba96c504917f35e3
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/71270
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>