analysis.os: bug 488757: Let the KernelTidAspect block if not queryable
This aspect implements the blocking resolve method and, if the caller
wishes, will block until the analysis is ready to be queried at the
time of the event. A progress monitor is used to allow for cancelling
the blocking call.

This patch updates the system call latency analysis to block.

Change-Id: Ie858bb4ed8984f6320af443030db4950ab5f7927
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/71604
Reviewed-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>