analysis.lami: Add support for user-defined LAMI analyses
This patch adds support for adding user-defined LAMI analyses,
as well as removing them, through the UI (project explorer).

A .properties file is created in the workspace for each added
user-defined analysis.

To add a new external analysis, right-click the External Analyses
item in the Project Explorer View, and click Add External Analysis.

To remove an external analysis, right-click the item to remove,
and click Remove External Analysis.

Change-Id: I067f0faf679d2384121a32845864d886ee64e241
Signed-off-by: Philippe Proulx <pproulx@efficios.com>
Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>
Reviewed-on: https://git.eclipse.org/r/72103
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-by: Hudson CI