analysis: Fix storing of fields of system call latency analysis
Right now all fields stored as map fieldName -> field value using
the toString() method on the field value (field.getValue().toString)
However, this cause a problem when the type the value is an object
other than string or a primitive type (e.g. long[]). This patches
changes this to uses field.toString() which creates a formatted string
of the value.

Change-Id: I6ff61eb70d7a7bb958b77797c3271c61cc430eef
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/71698
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>