xml: bug 493954: Allow string states to be displayed in time graph views
If a string state value is to be displayed in the time graph, it will add
a new state to the presentation provider with color. This state will be re-used
when the same string is displayed again. This is ideal for time graphs of state
system where there is a fixed number of strings as value, yet impossible to
define with definedValues because specific to a trace.

Change-Id: I2d1fa041e578e0a7dbccc3138153db3fd4684c62
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/71970
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>