analysis.lami: Replace OnDemandAnalysisException with CoreException
We may want to pass additional information to the exceptions thrown
when a LAMI analysis does not complete succesfully, like what
type of error should be displayed.

CoreException and IStatus are made exactly for that, and they hook
nicely into ErrorMessage to display, for example, a command-line
stderr output as additional Details in the dialog.

That way, a message like "no results returned" won't look like
a scary error.

Bug: 493941

Change-Id: Id3e6711ac410c3d993b4928a350dc1bdbcf89f5a
Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>
Reviewed-on: https://git.eclipse.org/r/72242
Reviewed-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>