ss: Replace AttributeNotFoundException with IOOBE for quark parameters
Methods that take a quark integer as parameter are changed to throw a
runtime IndexOutOfBoundsException instead of a checked
AttributeNotFoundException.

This makes those methods consistent with other methods that already
throw IOOBE when the quark is out of range.

Many occurrences were already swallowing the exception or re-throwing a
runtime exception.

Change-Id: I943b407e07bbc226c6b7c03306cd6d00f783bca2
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/72198
Reviewed-by: Alexandre Montplaisir <alexmonthy@efficios.com>
Reviewed-by: Hudson CI