ss: Improve attribute tree encoding to disk
The attribute tree used to be encoded as a list of full attribute
paths.

This patch improves the encoding by replacing common subpaths from
one attribute to the next by "*".

For example, if the previous attribute was "Threads/42/Status", and the
current attribute is "Threads/42/Prio", we encode the current
attribute as "*/*/Prio".

On common Kernel Analysis State Systems, the Attribute Tree is 1 MiB
per 50k attributes, with this patch the Attribute Tree size on disk
is reduced by 75%.

Change-Id: I53b119eba760c10b720e9a3c173ed2996c02c2a2
Signed-off-by: Loïc Prieur-Drevon <loic.prieurdrevon@gmail.com>
Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>
Reviewed-on: https://git.eclipse.org/r/72213
Reviewed-by: Hudson CI
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>