linux.ui: Add Operating System Overview Perspective
When solving a problem, one must locate it before they can
identify it.

The current LTTng Kernel perspective is very good at identifying
problems but its views are mostly useful when a user is zoomed in.

This perspective, allows users to locate problem using metrics that
they are used to: CPU usage, Memory usage and IO usage, just like
system monitors in Linux[1] and Windows[2].

This perspective is intended to be used in tandem with the LTTng
Kernel perspective as this one can hone in on an issue and the
Kernel perspective can figure out the problem.

[1] http://static.thegeekstuff.com/wp-content/uploads/2009/10/system-monitor-resources.png
[2] http://www.windows7library.com/blog/wp-content/uploads/2011/04/Resource-Monitor.jpg

Change-Id: I0465b5bd61dc6cca6e8b6f78adbf2969df6f5481
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/72303
Reviewed-by: Hudson CI
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Tested-by: Genevieve Bastien <gbastien+lttng@versatic.net>