critical path: bug 489360 Build the critical path in a separate thread
The critical path content provider was waiting for the analysis to be completed
but this often happens in the main thread and the first time, it is the whole
graph that needs to be computed and that can block the main thread for a while.

This patch uses a separate build thread when the input is changed so it does
not block anymore.

Change-Id: I9cb99c0d93524961c5f8b45ada59d8a02db49bb0
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/72759
Reviewed-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>