segStore: Add a close method to ISegmentStore
This method will close the segment store backend at the end of the analysis or
delete the files that were created if the analysis did not create correctly.

This method will be useful when there is multiple segment store backends, so
that each one can take care of its own saving actions and cleanup.

Change-Id: I7c2bd4c21b9141b0b4f0f3eed2495895299d3dc8
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/72931
Reviewed-by: Hudson CI
Reviewed-by: Jean-Christian Kouame <jean-christian.kouame@ericsson.com>
Tested-by: Jean-Christian Kouame <jean-christian.kouame@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>