rcp: Sign the RCP on Mac
The mac signing has to happen between materialize-products and
archive-products goals. Because we only want to do it when the
signing profile is enabled (eclipse.org server), we have to duplicate
some xml to fit the macsigner between the two. I did extract some
variables in order to make the duplication less error prone.

Bug: 488873
Change-Id: I65168e010dc866ef48b643f6aaf30aa94aedefda
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/72976
Reviewed-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>