critical path: bug 490086 Set start/end time for the trace
Previously, the start and end times of the critical path were set when building
the entries list for each critical path. Each worker had different start/end
times, different from the trace's and when getting a previously calculated
critical path, it would not reset the start/end time so the results were not
displayed correctly.

This patch sets the start and end time of the critical path view with the data
from the main execution graph, not the current worker's critical path.

It also gets the previously computed statistics in the content provider instead
of the analysis start.

Change-Id: I7c08f7776b375b033575bc56c51d1cff36b72068
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/73513
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Hudson CI