critical path: bug 490085 Always start critical path from head vertex
Previously, the critical path was calculated starting at the start of the
window range, but that behavior was not intuitive to the user and it was not
possible to change that window range once it is calculated.

Now the critical path is calculated for the whole of the worker's lifetime.

Change-Id: Id76192eec04dff6d3c45c1bd1314248d4ea991fb
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/73515
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
(cherry picked from commit 0039d5e8f603f10e079fe379f6767543a9e96786)
Reviewed-on: https://git.eclipse.org/r/74038
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>