tmf.xml: Bug 495210. Delete existing supplementary files
Previously, if we import a file that was previously loaded once, traces
won't be notified that the file changed.
Now, if the file already exist, the supplementary files for the analysis
are deleted and the related traces are closed.

Change-Id: I62ce6e6f4295db5f0c82e4c77ae73b4e5a4c1b85
Signed-off-by: Jean-Christian Kouame <jean-christian.kouame@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/73648
Tested-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Reviewed-by: Hudson CI
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>
(cherry picked from commit cfe92fc5a3135815d65af47a3128c108ff61b8b2)
Reviewed-on: https://git.eclipse.org/r/74296