analysis: Add totals to latency statistics view
Having the total is useful to know how much time was spent in a given
segment type.

For example, I have a Maven output analysis that computes the beginning
and end of each goal. By having the total, I can quickly see which type
of goal took the longest and then I can drill down to single instances.

Change-Id: Id59a81191726ad926821f7a250d56f4fbc622f7a
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/73770
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>