tmf: Bug 494952: Remove deadlock in Time Chart view
The handling of signal TmfTraceUpdated is now done in a
ProcessTraceThread. There is only one such thread per trace active at a
time. If an update is requested while a thread is active, the thread is
marked to be restarted after its current iteration completion.

The starting or restarting of a ProcessTraceThread is consolidated in a
method and synchronized.

The ProcessTraceThread now always processes the full time range of the
trace, even when the view is opened with a trace already opened.

A missing synchronization on DecorateThread concurrent access is added.

Method ITimeDataProvider.resetStartFinishTime(boolean) is added to allow
the TimeChartView to reset its window range without notifying listeners.

Change-Id: I0d49d2712af4e5c645bd61bfde46ba37fc42779c
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/73996
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>