tmf: Bug 495054: Importing or exporting custom parsers fails silently
When importing custom parsers from a file that does not contain any
custom parser definition, a message dialog is shown to the user.

When exporting a custom parser to an existing file that is not a valid
custom parser definition file, a message is logged to the Error Log.

Change-Id: I2f0bb781d9d43d8bbaf05da2b7ec06c2a65d47d3
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/74094
Reviewed-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
(cherry picked from commit d1b0903fe92fabef842f26e50cf12062d398c28c)
Reviewed-on: https://git.eclipse.org/r/74114