tmf: Bug 494810: SelectionEvent not sent after using vertical slider
The vertical slider used by the TmfVirtualTable is created with the
SWT.NO_FOCUS style. In Linux-GTK, this hint is not respected and the
slider can gain focus. When this happens, give back the focus to the
table.

This allows the SelectionEvent to be sent when extending a selection
using Shift+Click in the event table after using the vertical slider.

Change-Id: I9a8d7083300a5c515c24110d5572d30b6bd9d7d6
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/74119
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
(cherry picked from commit 2e200fc776de05e3928a2bc33df7a8b647b7d140)
Reviewed-on: https://git.eclipse.org/r/74475